Allow mixed arithmetic operators
The `no-mixed-operators` rules warns when mixing different operators in
the same expression without explicitly using parentheses.

While this is generally a good warning, it’s annoying with arithmetic
operations, because people generally understand the precedence of these
operators (`*` and `/` before `+` and `-`).  So, this PR updates the
`no-mixed-operators` rule to allow mixed arithmetic operators.  It
still warns about bit-wise, comparison, and logical operators.