Use 'off' and 'warn' in place of 0 and 1
A while back, ESLint added the ability to use the strings `off`,
`warn`, and `error` in place of the more opaque numeric levels (0, 1,
and 2).

This PR converts our configs to use the strings.  This makes our
configs easier to understand, particularly for those not terribly
familiar with the ESLint config file format.