Upgrade to eslint-find-rules 3.0
This version no longer reports deprecated rules as unused, so we can safely remove the `--no-error` flag from our `missing-rules` scripts.

In order to not fail the build, we have to lock our lint plugin `devDependencies` down to a specific version to make sure we're testing against the exact version we support.