🎨 Turn on rules to enforce sorting
The diff is rather large, because turning on sorting rules forced us to sort our rule configuration keys.

Here are the actual rule changes:

```js
    "sort-keys": ["warn", "asc", { natural: true }],
    "import/order": [
      "warn",
      {
        groups: [
          ["builtin", "external"],
          "internal",
          "parent",
          ["sibling", "index"],
        ],
        "newlines-between": "always",
      },
    ],
    "react/jsx-sort-props": [
      "warn",
      {
        callbacksLast: false,
        ignoreCase: true,
        shorthandFirst: false,
        shorthandLast: false,
      },
    ]
```

The `import/order` rule configuration is the best we can do for now.  We'd like a combination of that rule and some of what `sort-imports` does, but they conflict with each other.  There are open issues about adding a secondary alphabetical sort to `import/order`, but they haven't been moved forward in a while.