✨ Add eslint-plugin-jsx-a11y
This adds the jsx-a11y eslint plugin for use in React projects.  The rule configuration is essentially `jsx-a11y/recommended`, but with everything as warnings and not errors, as errors are generally too heavy-handed and keep us from trying something out quickly without worrying about fixing all of the lint warnings right away.