🚨 Remove obsolete eslint disable comments
In #158, we disabled the `import/no-unused-modules` rule and in #159, enabled the new `reportUnusedDisableDirectives` setting.

However, we had added eslint-disable comments to several files to disable the `import/no-unused-modules` rule for those files.

Those comments are no longer necessary since the rule is disabled, but we didn't catch that when we enabled the new setting in 159.

This PR removes the now-unnecessary comments.